// src/domain/usecases/update-hotel-usecase.interface.ts
import { Address } from '../entities/hotel.types'
import { HotelRepositoryData } from '../repositories/hotel-repository.interface'

export type UpdateHotelUseCaseInput = {
  id: string
  name: string
  address: Address
}

export interface UpdateHotelUseCaseInterface {
  execute: (input: UpdateHotelUseCaseInput) => Promise<HotelRepositoryData>
}